Moved the AsyncTask class
